build_scripts/requirements.txt
# Python Requirements for Application Document Generator
# Install with: pip install -r requirements.txt

# Core dependencies
PyYAML>=6.0.1              # YAML file processing
pathlib2>=2.3.7            # Enhanced path handling (for older Python versions)

# PDF generation (optional - for future implementation)
# weasyprint>=60.0          # HTML/CSS to PDF conversion
# markdown>=3.5.1           # Markdown processing
# pymdown-extensions>=10.0   # Extended markdown functionality

# CLI utilities
argparse                   # Command-line argument parsing (built-in for Python 3.2+)

# Development dependencies (optional)
# black>=23.0.0             # Code formatting
# flake8>=6.0.0            # Code linting
# pytest>=7.0.0            # Testing framework

# Documentation (optional)
# mkdocs>=1.5.0            # Documentation generator
# mkdocs-material>=9.0.0   # Material theme for MkDocs

# Future enhancements
# jinja2>=3.1.0            # Advanced templating
# python-docx>=1.1.0       # Word document generation
# reportlab>=4.0.0         # PDF generation alternative
